Pull charts referenced by providers k0s.extensions during helm-package
In order to support airgap customers will need to place all of the charts
a ManagedCluster may pull upon into a registry.  This adds support to our
'helm-package' Makefile to ensure we package up all of the charts our templates
reference so that they can be included in an airgap bundle.

Signed-off-by: Kyle Squizzato <ksquizzato@mirantis.com>